// NextJS Design JAM/day4/types/product.ts
export interface Product{
    _id: string;
    title?: string;
    _type?: "product";
    productImage?:{
        asset:{
            _ref: string;
            _type: "image";
        }
    };
    price: number;
    description?: string;
    slug : {
        _type : "slug";
        current :string;
    }
    
    tags?: [];
    dicountPercentage?: number;
    isNew?: boolean;
    inventory: number;
}